Fix wrong type for GO_Type enum
Prior save games used 2 Bytes instead of 1 (while 1 would have been enough)
Fixes #1328

+++ libs/s25main/gameTypes/GO_Type.h
@@ -21,7 +21,7 @@
 
 /// To be able to load old savegames and keep ids unique, please insert new
 /// items at the end of the list.
-enum class GO_Type : uint8_t
+enum class GO_Type : uint16_t
 {
     Unknown,
     Nothing,
